+++ lucene/src/test/org/apache/lucene/search/similarities/LMSimilarity.java
@@ -67,7 +67,7 @@ public abstract class LMSimilarity extends EasySimilarity {
 
   @Override
   protected void explain(Explanation expl, EasyStats stats, int doc,
-      float freq, int docLen) {
+      float freq, float docLen) {
     expl.addDetail(new Explanation(collectionModel.computeProbability(stats),
                                    "collection probability"));
   }
